test(juniper-diesel-rocket): use different fixture for each test
After some experiment, we avoid using the setup and tear-down method
of cleaning up the database after and before each test.

Currently, Rust has no good support for something like Python `Conftest.py`.
We have to do some acrobats to do setup-and-teardown.

Most of the time, the tests are failing, because the teardown is not executed if
the test function is panic.

The best workaround we have currently is just to use different fixture for each test
function.